pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.12"
dependencies = [
    "fastapi[all]>=0.116.0",
    "numpy>=2.3.1",
    "pillow>=11.3.0",
    "python-multipart>=0.0.20",
    "scikit-learn>=1.7.0",
    "scipy>=1.16.0",
    "uvicorn[standard]>=0.35.0",
]
